refactor: group neqsim fluid provider files in infrastructure
