+++ tests/libecalc/domain/process/fluid_stream/test_thermo_system.py
@@ -5,7 +5,7 @@ import pytest
 from ecalc_neqsim_wrapper.thermo import NeqsimFluid
 from libecalc.domain.process.value_objects.fluid_stream.fluid_model import EoSModel
 from libecalc.domain.process.value_objects.fluid_stream.process_conditions import ProcessConditions
-from libecalc.infrastructure.thermo_system_providers.neqsim_thermo_system import NeqSimThermoSystem
+from libecalc.infrastructure.neqsim_fluid_provider.neqsim_thermo_system import NeqSimThermoSystem
 
 
 class TestNeqSimThermoSystem:
